requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
paramiko>=3.4.0
netmiko>=4.3.0
python-nmap>=0.7.1
PyYAML>=6.0.1
